+++ packages/react-make-styles/etc/react-make-styles.api.md
@@ -7,20 +7,18 @@
 import { createDOMRenderer } from '@fluentui/make-styles';
 import type { CSSClassesMapBySlot } from '@fluentui/make-styles';
 import type { CSSRulesByBucket } from '@fluentui/make-styles';
-import { macros } from '@fluentui/make-styles';
 import type { MakeStaticStyles } from '@fluentui/make-styles';
 import type { MakeStylesRenderer } from '@fluentui/make-styles';
 import type { MakeStylesStyleRule } from '@fluentui/make-styles';
 import { mergeClasses } from '@fluentui/make-styles';
 import * as React_2 from 'react';
+import { shorthands } from '@fluentui/make-styles';
 import type { Theme } from '@fluentui/react-theme';
 
 // @internal
 export function __styles<Slots extends string>(classesMapBySlot: CSSClassesMapBySlot<Slots>, cssRules: CSSRulesByBucket): () => Record<Slots, string>;
 
 export { createDOMRenderer }
-
-export { macros }
 
 // @public (undocumented)
 export function makeStaticStyles<Selectors>(styles: MakeStaticStyles | MakeStaticStyles[]): () => void;
@@ -44,6 +42,8 @@ export interface RendererProviderProps {
 
 // @public
 export function renderToStyleElements(renderer: MakeStylesRenderer): React_2.ReactElement[];
+
+export { shorthands }
 
 // @public
 export function useRenderer(): MakeStylesRenderer;
